ACCUMULO-4086 got IT test working
